feat(features): use React Query to fetch document collection data
