+++ src/features/Viz/index.ts
@@ -1,3 +1,3 @@
 export { default as FeaturedObjViz } from './Viz';
 export type { TFamily, Collections } from './types';
-export { default as getData } from './hooks/getDocumentsData';
+export { default as useGetDocumentsCollection } from './hooks/useGetDocumentsCollection';
